Address review comments

+++ villagesql/bld_tools/add_non_root_docker_user.sh
@@ -8,26 +8,34 @@
 # GROUP_ID   - gid for the group the user would be part of
 # GROUP_NAME - group name for the group
 
-# Ubuntu images come prebuilt with user(ubuntu, 1000) and group(ubuntu, 1000). # Delete existing user(if any) on USER_ID.
+# User cannot be root
+if ((USER_ID == 0 || GROUP_ID == 0)) ||
+	[[ "$USER_NAME" == "root" || "$GROUP_NAME" == "root" ]]; then
+	echo "root user/group values are not allowed."
+	exit 1
+fi
+
+# Ubuntu images come prebuilt with user(ubuntu, 1000) and group(ubuntu, 1000).
+# Delete existing user(if any) on USER_ID.
 if getent passwd "${USER_ID}" >/dev/null; then
-  userdel "$(getent passwd "${USER_ID}" | cut -d: -f1)"
+	userdel "$(getent passwd "${USER_ID}" | cut -d: -f1)"
 fi
 
 # Delete existing group(if any) on GROUP_ID
 if getent group "${GROUP_ID}" >/dev/null; then
-  groupdel "$(getent group "${GROUP_ID}" | cut -d: -f1)"
+	groupdel "$(getent group "${GROUP_ID}" | cut -d: -f1)"
 fi
 
 # USER_NAME might already exist on a different uid other than USER_ID
 # Delete it if so.
 if getent passwd "${USER_NAME}" >/dev/null; then
-  userdel "${USER_NAME}"
+	userdel "${USER_NAME}"
 fi
 
 # GROUP_NAME might already exist on a different gid other than GROUP_ID
 # Delete it if so.
 if getent group "${GROUP_NAME}" >/dev/null; then
-  groupdel "${GROUP_NAME}"
+	groupdel "${GROUP_NAME}"
 fi
 
 # Add (GROUP_ID, GROUP_NAME)
